library/pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "promptops"
version = "0.1.0"
description = "Lightweight SDK and CLI for prompt evaluation and CI/CD"
authors = [{name = "PromptOps Team", email = "team@promptops.dev"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
]

[project.scripts]
promptops = "promptops.cli:main"
promptwizard = "promptops.cli:deprecated_main"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.12.0",
    "ruff>=0.1.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ["py311"]
